[PIDM-1765] avoid splitting massive iban api test in dedicated file
